+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # IAM Changelog
 
-## Version 3.4.1 (current stable)
+## Version 3.4.2 (current stable)
+* add new endpoint api/pegawais/v3/info/from-iam-token
+
+## Version 3.4.1
 * fix bugs class PosisiHelper
 
 ## Version 3.4.0